refactor: rename typo
